Template For Classes
This is the template for all the classes that everyone should follow while writing your methods. The starter code format should be the same across the team in order to factor our code together for the demo. This is commite to the master branch because all the methods specifc to the project are empty and waited to be completed.